pyproject.toml
[tool.poetry]
name = "itmo-events"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.2.2"
SQLAlchemy = "^1.4.42"
requests = "^2.28.1"
alembic = "^1.8.1"
pydantic = "^1.10.2"
config = "^0.5.1"
psycopg2 = "^2.9.5"
make = "^0.1.6.post2"
pytest = "^7.2.0"
starlette = "^0.21.0"
SQLAlchemy-Utils = "^0.38.3"
gunicorn = "^20.1.0"
python-dotenv = "^0.21.0"
PyJWT = {extras = ["crypto"], version = "^2.6.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
quart = "^0.18.3"
asyncpg = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
